refactor: align project with hexagonal architecture
